requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# automate.py: 3
Pillow == 5.0.0

# automate.py: 8
beautifulsoup4 == 4.6.0

# automate.py: 4
requests == 2.18.4

# automate.py: 1
selenium == 3.11.0
